+++ include/math/operations/Comparison.hpp
@@ -4,4 +4,4 @@
 #include <cstdint>
 #include <vector>
 
-std::strong_ordering compareDigits(const std::vector<uint32_t>& lhs, const std::vector<uint32_t>& rhs);
+std::strong_ordering compareDigits(const std::vector<uint64_t>& lhs, const std::vector<uint64_t>& rhs);
