+++ include/math/operations/Multiplication.hpp
@@ -4,12 +4,12 @@
 #include <utility>
 #include <vector>
 
-std::vector<uint32_t> multiplyDigits(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand);
+std::vector<uint64_t> multiplyDigits(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand);
 
-std::vector<uint32_t> hardwareMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand);
-std::vector<uint32_t> longMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand);
-std::vector<uint32_t> karatsubaMultiplication(const std::vector<uint32_t>& multiplier, const std::vector<uint32_t>& multiplicand);
+std::vector<uint64_t> hardwareMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand);
+std::vector<uint64_t> longMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand);
+std::vector<uint64_t> karatsubaMultiplication(const std::vector<uint64_t>& multiplier, const std::vector<uint64_t>& multiplicand);
 
-std::pair<std::vector<uint32_t>, std::vector<uint32_t>> splitVector(const std::vector<uint32_t>& vec, size_t index);
+std::pair<std::vector<uint64_t>, std::vector<uint64_t>> splitVector(const std::vector<uint64_t>& vec, size_t index);
 
-std::vector<uint32_t> leftShiftDigits(const std::vector<uint32_t>& input, size_t shift);
+std::vector<uint64_t> leftShiftDigits(const std::vector<uint64_t>& input, size_t shift);
