+++ src/math/operations/Subtraction.cpp
@@ -1,8 +1,8 @@
 #include "math/operations/Subtraction.hpp"
 
-std::vector<uint32_t> subtractDigits(const std::vector<uint32_t>& minuend, const std::vector<uint32_t>& subtrahend)
+std::vector<uint64_t> subtractDigits(const std::vector<uint64_t>& minuend, const std::vector<uint64_t>& subtrahend)
 {
-	std::vector<uint32_t> difference;
+	std::vector<uint64_t> difference;
 	difference.reserve(minuend.size());
 
 	uint64_t borrow = 0;
@@ -11,7 +11,7 @@ std::vector<uint32_t> subtractDigits(const std::vector<uint32_t>& minuend, const
 		uint64_t diff = minuend[i] - borrow;
 		if (i < subtrahend.size()) diff -= subtrahend[i];
 
-		difference.emplace_back(static_cast<uint32_t>(diff));
+		difference.emplace_back(static_cast<uint64_t>(diff));
 		borrow = (diff >> 32) > 0;
 	}
